moves helpers folder to root
