+++ app/helpers/hash.js
